Reading enchancement part1
-improved methods for registering, saving and listing readings

+++ Kamienica/src/main/java/kamienica/feature/invoice/InvoiceWaterDAOImpl.java
@@ -11,7 +11,7 @@ import kamienica.feature.payment.PaymentStatus;
 
 @Repository("invoiceWater")
 @Transactional
-public class InvoiceWaterDAOImpl extends AbstractDao<Integer, InvoiceWater> implements InvoiceDao<InvoiceWater> {
+public class InvoiceWaterDAOImpl extends AbstractDao<Long, InvoiceWater> implements InvoiceDao<InvoiceWater> {
 
 	@Override
 	public InvoiceWater getLatest() {
